refactor: merge SessionTracker and SessionManager, convert BrowserDetector to functional approach
