+++ packages/agent/src/index.ts
@@ -12,14 +12,13 @@ export * from './tools/shell/listShells.js';
 export * from './tools/shell/ShellTracker.js';
 
 // Tools - Browser
-export * from './tools/session/lib/SessionManager.js';
 export * from './tools/session/lib/types.js';
 export * from './tools/session/sessionMessage.js';
 export * from './tools/session/sessionStart.js';
 export * from './tools/session/lib/PageController.js';
-export * from './tools/session/lib/BrowserAutomation.js';
 export * from './tools/session/listSessions.js';
 export * from './tools/session/SessionTracker.js';
+// Export browser detector functions
 
 export * from './tools/agent/AgentTracker.js';
 // Tools - Interaction
